Move CacheCourseFiles to RomFS class
Moved CacheCourseFiles out of Program.cs and into RomFS to run when SetRoot is called.

When romfs root is set with SetRoot the programmer can expect a populated list of course entries with RomFS.GetCourseEntries()
